feat: add flyway health check
